+++ tests/test_settings.py
@@ -1,5 +1,3 @@
-import django
-
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.sqlite3',
@@ -26,15 +24,10 @@ INSTALLED_APPS = [
     'request',
 ]
 
-MIDDLEWARE_CLASSES = [
+MIDDLEWARE = [
     'django.contrib.sessions.middleware.SessionMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
 ]
-
-# Old-style middleware using settings.MIDDLEWARE_CLASSES is deprecated since django 1.10
-if django.VERSION >= (1, 10):
-    MIDDLEWARE = MIDDLEWARE_CLASSES
-    del MIDDLEWARE_CLASSES
 
 STATIC_URL = '/static/'